+++ engine/MapReslistGenerator.h
@@ -28,7 +28,7 @@ struct maplist_map_t
 //	pMapFile: If you're using a maplist file, this should be the filename of it
 //	pSystemMsg: Used to preface and debug messages
 //	iCurrentMap: The map in the list to begin at. Handles the -startmap parameter for you.
-bool BuildGeneralMapList( CUtlVector<maplist_map_t> *aMaps, bool bUseMapListFile, const char *pMapFile, const char *pSystemMsg, int *iCurrentMap );
+bool BuildGeneralMapList( CUtlVector<maplist_map_t> *aMaps, bool bUseMapListFile, const char *pMapFile, const char *pSystemMsg, intp *iCurrentMap );
 
 // initialization
 void MapReslistGenerator_Init();
@@ -102,7 +102,7 @@ private:
 
 	// list of all maps to scan
 	CUtlVector<maplist_map_t> m_Maps;
-	int m_iCurrentMap;
+	intp m_iCurrentMap;
 	float m_flNextMapRunTime;
 	int m_iFrameCountdownToRunningNextMap;
 	CUtlSymbolTable m_AlreadyWrittenFileNames;
